tidied up the namespace modules
