+++ frontend/src/app/screens/mod.rs
@@ -1,0 +1,1 @@
+// TODO adjust all paths with new ones in documentation that reference old ones
