+++ frontend/src/screens.rs
@@ -1,0 +1,24 @@
+pub mod articles_screen;
+pub use articles_screen::ArticlesScreen;
+pub mod example_screen;
+pub use example_screen::ExampleScreen;
+pub mod game;
+pub use game::Game;
+pub mod game_setup_screen;
+pub use game_setup_screen::GameSetupScreen;
+pub mod lobby_setup;
+pub use lobby_setup::LobbySelectionScreen;
+pub mod main_menu;
+pub use main_menu::MainMenu;
+pub mod mcg_lobby;
+pub use mcg_lobby::LobbyScreen;
+pub mod pairing_screen;
+pub use pairing_screen::PairingScreen;
+pub mod poker;
+pub use poker::screen::PokerOnlineScreen;
+pub mod qr_test;
+pub use qr_test::QrScreen;
+pub mod qr_test_receive;
+pub use qr_test_receive::QrTestReceive;
+pub mod qr_test_transmit;
+pub use qr_test_transmit::QrTestTransmit;
