+++ frontend/src/screens/poker/player_manager.rs
