+++ frontend/src/widgets/card.rs
