+++ frontend/src/widgets/mod.rs
@@ -1,0 +1,6 @@
+pub mod screen;
+pub mod card;
+pub mod hardcoded_cards;
+pub mod field;
+pub mod qr_scanner;
+pub mod theme;
