+++ frontend/src/widgets/qr_scanner.rs
